Add complete Push support
Still no remote delete (may not add it at all as I don't think I'll be needing it).

Important things to remember:

coreResource.resourceId should be nil for non-remotely inserted objects.
Any object that will be remotely synchronized should have a resourceId.

Changes:
[CoreManager managedObjectContext] is now thread aware.
[CoreManager newManagedObjectContext] passes an autoreleased object that should be used only on the current thread.

CoreManager now listens for NSManagedObjectContextDidSaveNotification so that an assertion error will be called if a managedobjectcontext is saved on a non-main thread without merging changes. Turn this off in production code.

Removed a bunch of logging stuff using NSLog and replaced it with DLog (see the .pch file for the macro).